Download and register DAC with CLRMD
Issue https://github.com/dotnet/diagnostics/issues/158

Add SOSInitializeByHost SOS module entry so the SOSHost
code can pass the temp dir created, DAC module path downloaded for
clrmd and the symbol store enabled flag.

Removed ISOSHostServices and replaced initializing the SOSNetCoreCallbacks
initialization as part of the call to SOSInitializeByHost.

Added the "clrmodules" command as a test of a pure CLRMD command and
as place to add the module version requested by some customers.

Catch the DllNotFoundException exception and add "try installing libc6-dev"
to exception message and documentation to work around Microsoft stretch docker
image problems.

Issue: https://github.com/dotnet/diagnostics/issues/201
