+++ documentation/dotnet-dump-instructions.md
@@ -5,7 +5,7 @@ The dotnet-dump CLI global tool is way to collect and analyze Windows and Linux 
 
 ## Installing dotnet-dump
 
-The first step is to install the dotnet-dump CLI global tool. This requires the 2.1 .NET Core SDK to be installed. If you see the error message `Tool 'dotnet-dump' is already installed`, you will need to uninstall the global tool (see below). 
+The first step is to install the dotnet-dump CLI global tool. This requires at least the 2.1 or greater .NET Core SDK to be installed. If you see the error message `Tool 'dotnet-dump' is already installed`, you will need to uninstall the global tool (see below). 
 
     $ dotnet tool install -g dotnet-dump --version 1.0.3-preview5.19251.2
     You can invoke the tool using the following command: dotnet-dump
