+++ src/SOS/Strike/Strike.vcxproj.filters
@@ -57,7 +57,6 @@
       <Filter>inc</Filter>
     </ClInclude>
     <ClInclude Include="hostcoreclr.h" />
-    <ClInclude Include="soshostservices.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="Native.rc" />
